pyproject.toml
[project]
name = "mcp-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "langchain-mcp-adapters>=0.1.13",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
